This code is ECG device dependent. The ECM check only runs if the ECG device is properly attached and configured.
To run the program, double click rxboxGUIDemo.py

Added on the Icons folder are the different bitmaps which updates the status of each ECG electrode. 

Initially, the electrode placement bitmap shows the actual color and labels as seen on the actual electrodes.
When the play button is pressed the bitmap is updated, if the electrode has contact, it changes its color to green otherwise it changes to red. Lastly, if the stop button is pressed it returns to its initial image.